Time: 15 ms (27.89%) | Memory: 37.4 MB (38.58%) - LeetSync

+++ 2090-number-of-ways-to-arrive-at-destination/number-of-ways-to-arrive-at-destination.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int countPaths(int n, vector<vector<int>>& roads) {
-        int mod = (int)(1e9+7);
+        int mod = 1e9+7;
         vector<vector<pair<int,int>>> adj(n);
         for(auto it : roads){
             int u = it[0];
